split api endpoints into separate routers
